Resolved continuations and var statement declarations
Use vars when you need them. It's more readable and that block of code becomes more understandable. If we declare every variable at the top of the function and then initialize them when we need them, there may be confusion between the lines of the declaration and the lines of the initialization.